completed logout for frontend

+++ client/src/components/MyNavbar.js
@@ -7,14 +7,29 @@ import { Search } from "react-bootstrap-icons";
 import { NavLink } from "react-router-dom";
 
 const MyNavbar = () => {
-	// Client-Side logout function
 	async function logout() {
-		// Make a request to server to invalidate the JWT token
+		// retrieve token from session storage
+		const token = sessionStorage.getItem("token");
+
+		// Make a request to the server to invalidate the JWT token
 		try {
-			// Clear all items from sessionStorage
-			sessionStorage.clear();
+			const response = await fetch("http://localhost:3050/account/logout", {
+				method: "POST",
+				headers: {
+					Authorization: token ? token : "",
+				},
+			});
+
+			if (response.ok) {
+				// Clear all items from sessionStorage
+				sessionStorage.clear();
+			} else {
+				console.error("Error during logout:", response.statusText);
+				// Handle the error, display a message, or perform other actions
+			}
 		} catch (error) {
 			console.error("Error during logout:", error);
+			// Handle the error, display a message, or perform other actions
 		}
 	}
